backend/requirements.txt
fastapi==0.128.2
uvicorn[standard]==0.40.0
python-multipart==0.0.22
ffmpeg-python==0.2.0
google-cloud-storage
google-auth
